api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gspread==6.1.2
google-auth==2.35.0
pydantic==2.9.2
